lcw
订单详情页加点击查看单品详情

+++ TiJian/TiJian/ViewControllers/个人定制/PersonalCustomViewController.m
@@ -1129,7 +1129,7 @@
         [self addSortArrayWithQuestionId:questionId extension:YES];
     }
     
-    _questionId = [NSString stringWithFormat:@"%@_%ld",Q_Extension,questionId];
+    _questionId = [NSString stringWithFormat:@"%@_%ld",Q_Extension,(long)questionId];
     
     DDLOG(@"extension questionId: %@",_questionId);
